Refactor: migrate to uv & seperate example app
See https://github.com/edihasaj/tuspy-fast-api/issues/56

(or https://github.com/edihasaj/tuspyserver/issues/56 in the future)
