+++ examples/basic.py
@@ -1,10 +1,15 @@
+from tuspyserver import create_tus_router
+
 from fastapi import FastAPI
-from starlette.middleware.cors import CORSMiddleware
-from starlette.staticfiles import StaticFiles
+from fastapi.middleware.cors import CORSMiddleware
+from fastapi.staticfiles import StaticFiles
 
-from tusserver.tus import create_api_router
+import uvicorn
 
+# initialize a FastAPI app
 app = FastAPI()
+
+# configure cross-origin middleware
 app.add_middleware(
     CORSMiddleware,
     allow_origins=["*"],
@@ -21,25 +26,31 @@ app.add_middleware(
     ],
 )
 
+# serve an html frontend from the static folder
 app.mount("/static", StaticFiles(directory="static"), name="static")
 
 
+# use completion hook to log uploads
 def on_upload_complete(file_path: str, metadata: dict):
     print("Upload complete")
     print(file_path)
     print(metadata)
 
 
-def on_your_specific_auth():
-    pass
-
-
+# mount the tus router to our
 app.include_router(
-    create_api_router(
-        files_dir="/tmp/different_dir",
+    create_tus_router(
+        files_dir="./uploads",
         max_size=128849018880,
         on_upload_complete=on_upload_complete,
-        auth=on_your_specific_auth,
         prefix="files",
     )
 )
+
+# run the app with uvicorn
+if __name__ == "__main__":
+    uvicorn.run(
+        "basic:app",
+        reload=True,
+        use_colors=True,
+    )
