+++ examples/static/index.html
@@ -1,0 +1,47 @@
+<!doctype html>
+<html>
+
+<head>
+  <meta charset="utf-8">
+  <title>Uppy</title>
+  <link href="https://releases.transloadit.com/uppy/v3.3.1/uppy.min.css" rel="stylesheet">
+</head>
+
+<body>
+  <h1>tuspyserver example</h1>
+  <p>Uses <a href="https://uppy.io/docs/tus/">uppy</a> which uses <a
+      href="https://github.com/tus/tus-js-client">tus-js-client</a> under the hood.</p>
+  <div id="drag-drop-area"></div>
+
+  <script type="module">
+    import {
+      Uppy, Dashboard, Tus
+    } from "https://releases.transloadit.com/uppy/v3.3.1/uppy.min.mjs"
+    const uppy = new Uppy()
+      .use(Dashboard, {
+        inline: true,
+        target: '#drag-drop-area'
+      })
+      .use(Tus, {
+        endpoint: 'http://127.0.0.1:8000/files/'
+      })
+
+    uppy.on('complete', (result) => {
+      console.log('Upload complete! Files uploaded:', result.successful)
+    })
+  </script>
+</body>
+
+<style>
+  html {
+    display: grid;
+    place-items: center;
+  }
+
+  * {
+    text-align: center;
+    font-family: sans serif;
+  }
+</style>
+
+</html>
