--- requirements.txt
@@ -1,11 +1,0 @@
-fastapi~=0.110.0
-setuptools~=65.6.3
-anyio==4.3.0
-exceptiongroup==1.2.0
-click==8.1.7
-h11==0.14.0
-idna==3.6
-pydantic==2.6.2
-uvicorn==0.22.0
-sniffio==1.3.1
-typing_extensions==4.10.0
