--- static/uppy.html
@@ -1,25 +1,0 @@
-<!doctype html>
-<html>
-<head>
-    <meta charset="utf-8">
-    <title>Uppy</title>
-    <link href="https://releases.transloadit.com/uppy/v3.3.1/uppy.min.css" rel="stylesheet">
-</head>
-<body>
-<div id="drag-drop-area"></div>
-
-<script type="module">
-    import {Uppy, Dashboard, Tus} from "https://releases.transloadit.com/uppy/v3.3.1/uppy.min.mjs"
-    var uppy = new Uppy()
-        .use(Dashboard, {
-            inline: true,
-            target: '#drag-drop-area'
-        })
-        .use(Tus, {endpoint: 'http://127.0.0.1:8000/files/'})
-
-    uppy.on('complete', (result) => {
-        console.log('Upload complete! We’ve uploaded these files:', result.successful)
-    })
-</script>
-</body>
-</html>
